"Refactor block registration and downgrade plugin version"
This commit includes a significant refactor of the block registration method in DeferredRegistryHelper.java to streamline it and eliminate redundant queries to registries. This change will improve performance and readability. A TitaniumTab parameter was added to include the item in the tab list in case it exists.

In addition, Gradle properties and build files were edited. The api_version was bumped from 3.8.0 to 3.8.1. Also, the version of the 'se.bjurr.gitchangelog.git-changelog-gradle-plugin' was downgraded from 1.71.9 to 1.64 due to compatibility issues with the new api_version.